tests are actually working...close #24

+++ tests/test.sh
@@ -46,59 +46,59 @@ grabseqs -h
 #####
 
 ## test sample listing, metadata download
-#if [ `grabseqs sra -m SRP057027.tsv -l -o $TMPDIR/test_metadata/ SRP057027 | wc -l` -ne 370 ]; then
-#    exit 1
-#fi
-#echo -e "$PASS SRA sample listing test passed"
+if [ `grabseqs sra -m SRP057027.tsv -l -o $TMPDIR/test_metadata/ SRP057027 | wc -l` -ne 370 ]; then
+    exit 1
+fi
+echo -e "$PASS SRA sample listing test passed"
 
 ## test metadata download
-#if [ `cat $TMPDIR/test_metadata/SRP057027.tsv | wc -l` -ne 370 ] ; then
-#    exit 1
-#fi
-#echo -e "$PASS SRA metadata test passed"
+if [ `cat $TMPDIR/test_metadata/SRP057027.tsv | wc -l` -ne 370 ] ; then
+    exit 1
+fi
+echo -e "$PASS SRA metadata test passed"
 
 ## test behavior with -l and --no_parsing
-#if [ `grabseqs sra -l --no_parsing SRR1804203 | wc -l` -ne 1 ]; then
-#    exit 1
-#fi
-#echo -e "$PASS SRA sample listing test with SRR parsing disabled passed"
-#
+if [ `grabseqs sra -l --no_parsing SRR1804203 | wc -l` -ne 1 ]; then
+    exit 1
+fi
+echo -e "$PASS SRA sample listing test with SRR parsing disabled passed"
+
 ## unpaired fasterq-dump
-#grabseqs sra -t 2 -o $TMPDIR/test_tiny_sra ERR2279063
-#ls $TMPDIR/test_tiny_sra/ERR2279063.fastq.gz > /dev/null
-#echo -e "$PASS SRA unpaired sample download test passed"
+grabseqs sra -t 2 -o $TMPDIR/test_tiny_sra ERR2279063
+ls $TMPDIR/test_tiny_sra/ERR2279063.fastq.gz > /dev/null
+echo -e "$PASS SRA unpaired sample download test passed"
 
 ## paired fasterq-dump
-#grabseqs sra -t 2 -o $TMPDIR/test_tiny_sra_paired SRR1913936
-#ls $TMPDIR/test_tiny_sra_paired/SRR1913936_1.fastq.gz > /dev/null
-#ls $TMPDIR/test_tiny_sra_paired/SRR1913936_2.fastq.gz > /dev/null
-#echo -e "$PASS SRA paired sample download test passed"
-#
+grabseqs sra -t 2 -o $TMPDIR/test_tiny_sra_paired SRR1913936
+ls $TMPDIR/test_tiny_sra_paired/SRR1913936_1.fastq.gz > /dev/null
+ls $TMPDIR/test_tiny_sra_paired/SRR1913936_2.fastq.gz > /dev/null
+echo -e "$PASS SRA paired sample download test passed"
+
 ## unpaired fastq-dump
-#grabseqs sra -t 2 -o $TMPDIR/test_fastqdump_sra --use_fastq_dump ERR2279063
-#ls $TMPDIR/test_fastqdump_sra/ERR2279063.fastq.gz > /dev/null
-#echo -e "$PASS SRA unpaired sample download using fastq-dump test passed"
+grabseqs sra -t 2 -o $TMPDIR/test_fastqdump_sra --use_fastq_dump ERR2279063
+ls $TMPDIR/test_fastqdump_sra/ERR2279063.fastq.gz > /dev/null
+echo -e "$PASS SRA unpaired sample download using fastq-dump test passed"
 
 ## paired fasterq-dump
-#grabseqs sra -t 2 -o $TMPDIR/test_fastqdump_sra_paired --use_fastq_dump SRR1913936
-#ls $TMPDIR/test_fastqdump_sra_paired/SRR1913936_1.fastq.gz > /dev/null
-#ls $TMPDIR/test_fastqdump_sra_paired/SRR1913936_2.fastq.gz > /dev/null
-#echo -e "$PASS SRA paired sample download using fastq-dump test passed"
+grabseqs sra -t 2 -o $TMPDIR/test_fastqdump_sra_paired --use_fastq_dump SRR1913936
+ls $TMPDIR/test_fastqdump_sra_paired/SRR1913936_1.fastq.gz > /dev/null
+ls $TMPDIR/test_fastqdump_sra_paired/SRR1913936_2.fastq.gz > /dev/null
+echo -e "$PASS SRA paired sample download using fastq-dump test passed"
 
 ## test no clobber
-#t=`grabseqs sra -t 2 -o $TMPDIR/test_fastqdump_sra ERR2279063`
-#echo $t
-#if [[ $t != *"Pass -f to force download"* ]] ; then
-#    exit 1
-#fi
-#echo -e "$PASS SRA no-clobber test passed"
+t=`grabseqs sra -t 2 -o $TMPDIR/test_fastqdump_sra ERR2279063`
+echo $t
+if [[ $t != *"Pass -f to force download"* ]] ; then
+    exit 1
+fi
+echo -e "$PASS SRA no-clobber test passed"
 
 ## test force
-#tf=`grabseqs sra -t 2 -o $TMPDIR/test_fastqdump_sra -f ERR2279063`
-#if [[ $tf == *"Pass -f to force download"* ]] ; then
-#    exit 1
-#fi
-#echo -e "$PASS SRA force download test passed"
+tf=`grabseqs sra -t 2 -o $TMPDIR/test_fastqdump_sra -f ERR2279063`
+if [[ $tf == *"Pass -f to force download"* ]] ; then
+    exit 1
+fi
+echo -e "$PASS SRA force download test passed"
 
 ##########
 # iMicrobe
@@ -132,7 +132,7 @@ echo -e "$PASS iMicrobe fasta-formatted sample download test passed"
 grabseqs imicrobe -o $TMPDIR/test_tiny_im s6399
 ls $TMPDIR/test_tiny_im/s6399_1.fastq.gz  > /dev/null
 ls $TMPDIR/test_tiny_im/s6399_2.fastq.gz  > /dev/null
-echo -e "PASS iMicrobe fastq-formatted sample download test passed"
+echo -e "$PASS iMicrobe fastq-formatted sample download test passed"
 
 ## test no clobber
 t=`grabseqs imicrobe -t 2 -o $TMPDIR/test_tiny_im s710`
@@ -154,41 +154,41 @@ echo -e "$PASS iMicrobe force download test passed"
 #########
 
 ## test sample listing, metadata download
-#if [ `grabseqs mgrast -o $TMPDIR/test_md_mg -m META.csv -l mgp85479 | wc -l` -ne 5 ]; then
-#    exit 1
-#fi
-#echo -e "$PASS MG-RAST sample listing test passed"
+if [ `grabseqs mgrast -o $TMPDIR/test_md_mg -m META.csv -l mgp85479 | wc -l` -ne 5 ]; then
+    exit 1
+fi
+echo -e "$PASS MG-RAST sample listing test passed"
 
 ## test metadata
-#if [ `cat $TMPDIR/test_md_mg/META.csv | wc -l` -ne 5 ] ; then
-#    exit 1
-#fi
-#echo -e "$PASS MG-RAST metadata test passed"
+if [ `cat $TMPDIR/test_md_mg/META.csv | wc -l` -ne 5 ] ; then
+    exit 1
+fi
+echo -e "$PASS MG-RAST metadata test passed"
 
 ## download a tiny sample, .fastq-formatted
-#grabseqs mgrast -o $TMPDIR/test_tiny_mg mgm4793571.3
-#ls $TMPDIR/test_tiny_mg/mgm4793571.3.fastq.gz > /dev/null
-#echo -e "$PASS MG-RAST unpaired sample download test passed (fastq-formatted)"
+grabseqs mgrast -o $TMPDIR/test_tiny_mg mgm4793571.3
+ls $TMPDIR/test_tiny_mg/mgm4793571.3.fastq.gz > /dev/null
+echo -e "$PASS MG-RAST unpaired sample download test passed (fastq-formatted)"
 
 ## download a tiny sample, .fasta-formatted
-#grabseqs mgrast -o $TMPDIR/test_tiny_mg_fasta mgm4633450.3
-#ls $TMPDIR/test_tiny_mg_fasta/mgm4633450.3.fastq.gz > /dev/null
-#echo -e "$PASS MG-RAST fasta-formatted sample download test passed"
+grabseqs mgrast -o $TMPDIR/test_tiny_mg_fasta mgm4633450.3
+ls $TMPDIR/test_tiny_mg_fasta/mgm4633450.3.fastq.gz > /dev/null
+echo -e "$PASS MG-RAST fasta-formatted sample download test passed"
 
 ## test no clobber
-#u=`grabseqs mgrast -o $TMPDIR/test_tiny_mg mgm4793571.3`
-#echo $u
-#if [[ $u != *"Pass -f to force download"* ]] ; then
-#    exit 1
-#fi
-#echo -e "$PASS MG-RAST no-clobber test passed"
+u=`grabseqs mgrast -o $TMPDIR/test_tiny_mg mgm4793571.3`
+echo $u
+if [[ $u != *"Pass -f to force download"* ]] ; then
+    exit 1
+fi
+echo -e "$PASS MG-RAST no-clobber test passed"
 
 ## test force
-#u=`grabseqs mgrast -o $TMPDIR/test_tiny_mg -f mgm4793571.3`
-#if [[ $u == *"Pass -f to force download"* ]] ; then
-#    exit 1
-#fi
-#echo -e "$PASS MG-RAST force download test passed"
+u=`grabseqs mgrast -o $TMPDIR/test_tiny_mg -f mgm4793571.3`
+if [[ $u == *"Pass -f to force download"* ]] ; then
+    exit 1
+fi
+echo -e "$PASS MG-RAST force download test passed"
 
 ## test conda install
 conda install -c louiejtaylor -qy grabseqs
